Extract the StreamSpliterator for reuse and mark it @Internal
Extract the StreamSpliterator for reuse and mark it @Internal

#3102
